+++ android/src/main/java/com/mackentoch/beaconsandroid/BeaconsAndroidBootReceiver.java
@@ -3,6 +3,7 @@ package com.mackentoch.beaconsandroid;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
+import android.os.Build;
 import android.util.Log;
 
 import static com.mackentoch.beaconsandroid.BeaconsAndroidModule.LOG_TAG;
@@ -14,6 +15,12 @@ public class BeaconsAndroidBootReceiver extends BroadcastReceiver
 	{
 		Log.d(LOG_TAG, "onReceive...");
 
-		context.startService(new Intent(context, BeaconsAndroidTransitionService.class));
+		//context.startService(new Intent(context, BeaconsAndroidTransitionService.class));
+
+		if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+			context.startForegroundService(new Intent(context, BeaconsAndroidTransitionService.class));
+		} else {
+			context.startService(new Intent(context, BeaconsAndroidTransitionService.class));
+		}
 	}
 }